requirements.txt
GitPython
yaml